refactor: MacBook 역직렬화 클래스 구현
