--- src/main/kotlin/backend/itracker/crawl/domain/MacBookCategory.kt
@@ -1,8 +1,0 @@
-package backend.itracker.crawl.domain
-
-enum class MacBookCategory (
-    val korean: String
-){
-    MACBOOK_AIR("맥북 에어"),
-    MACBOOK_PRO("맥북 프로"),
-}
