+++ src/main/kotlin/backend/itracker/crawl/service/macbook/MacBookCrawlMapper.kt
@@ -1,0 +1,4 @@
+package backend.itracker.crawl.service.macbook
+
+class MacBookCrawlMapper {
+}
